refactor, split preview-model from preview...
